requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.3.2
langchain==0.1.9
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.0 